MatrixPilot_beta: commit branch updates to repository
git-svn-id: https://gentlenav.googlecode.com/svn/branches/MatrixPilot_beta@3391 cf4e4d4a-e611-11dd-8fb1-9960f2a117f8

+++ MAVLink/include/checksum.h
@@ -56,12 +56,12 @@ static inline void crc_init(uint16_t* crcAccum)
  **/
 static inline uint16_t crc_calculate(const uint8_t* pBuffer, uint16_t length)
 {
-        uint16_t crcTmp;
-        crc_init(&crcTmp);
+    uint16_t crcTmp;
+    crc_init(&crcTmp);
 	while (length--) {
-                crc_accumulate(*pBuffer++, &crcTmp);
-        }
-        return crcTmp;
+        crc_accumulate(*pBuffer++, &crcTmp);
+    }
+    return crcTmp;
 }
 
 /**
@@ -73,7 +73,7 @@ static inline uint16_t crc_calculate(const uint8_t* pBuffer, uint16_t length)
  * @param data new bytes to hash
  * @param crcAccum the already accumulated checksum
  **/
-static inline void crc_accumulate_buffer(uint16_t *crcAccum, const char *pBuffer, uint8_t length)
+static inline void crc_accumulate_buffer(uint16_t *crcAccum, const char *pBuffer, uint16_t length)
 {
 	const uint8_t *p = (const uint8_t *)pBuffer;
 	while (length--) {
